+++ onchain/rollups/deploy/02_factory.ts
@@ -23,6 +23,7 @@ const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
 
     const { Bitmask, MerkleV2 } = await deployments.all();
 
+    await deployments.deploy("QuorumFactory", opts);
     await deployments.deploy("CartesiDAppFactory", {
         ...opts,
         libraries: {